chore: Move update model endpoint to endpoints section of app
