+++ tests/endpoints/tests_models.py
@@ -3,6 +3,7 @@ from unittest.mock import patch
 from canonicalwebteam.candid import CandidClient
 from canonicalwebteam.exceptions import (
     StoreApiResponseErrorList,
+    StoreApiResourceNotFound,
 )
 from webapp.helpers import api_publisher_session
 from tests.endpoints.endpoint_testing import TestModelServiceEndpoints
@@ -71,3 +72,49 @@ class TestGetModels(TestModelServiceEndpoints):
         self.assertEqual(response.status_code, 500)
         self.assertFalse(data["success"])
         self.assertIn("Store not found", data["message"])
+
+
+class TestUpdateModel(TestModelServiceEndpoints):
+    @patch(
+        "canonicalwebteam.store_api.publishergw.PublisherGW.update_store_model"
+    )
+    def test_update_model(self, mock_update_store_model):
+        mock_update_store_model.return_value = None
+
+        payload = {"api_key": self.api_key}
+        response = self.client.patch(
+            "/api/store/1/models/Model1", data=payload
+        )
+        data = response.json
+
+        self.assertEqual(response.status_code, 200)
+        self.assertTrue(data["success"])
+
+    def test_update_model_with_invalid_api_key(self):
+        payload = {"api_key": "invalid_api_key"}
+        response = self.client.patch(
+            "/api/store/1/models/Model1", data=payload
+        )
+        data = response.json
+
+        self.assertEqual(response.status_code, 500)
+        self.assertFalse(data["success"])
+        self.assertEqual(data["message"], "Invalid API key")
+
+    @patch(
+        "canonicalwebteam.store_api.publishergw.PublisherGW.update_store_model"
+    )
+    def test_model_not_found(self, mock_update_store_model):
+        mock_update_store_model.side_effect = StoreApiResourceNotFound(
+            "Model not found", 404, [{"message": "Model not found"}]
+        )
+
+        payload = {"api_key": self.api_key}
+        response = self.client.patch(
+            "/api/store/1/models/Model1", data=payload
+        )
+        data = response.json
+
+        self.assertEqual(response.status_code, 500)
+        self.assertFalse(data["success"])
+        self.assertEqual(data["message"], "Model not found")
